+++ packages/editor/editor-core/src/plugins/table/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './selection';
-export * from './position';
 export * from './decoration';
 export * from './nodes';