+++ website/src/containers/Nav/index.js
@@ -1,8 +1,8 @@
-/* @flow */
+// @flow
 
 import React, { Component } from 'react';
 import styled from 'styled-components';
-import { Link, Switch, Route, matchPath } from 'react-router-dom';
+import { Link, Switch, Route } from 'react-router-dom';
 import Navigation, {
   AkContainerTitle,
   presetThemes,
@@ -12,21 +12,14 @@ import Tooltip from '@atlaskit/tooltip';
 
 import SearchIcon from '@atlaskit/icon/glyph/search';
 import MenuIcon from '@atlaskit/icon/glyph/menu';
-import AtlassianIcon from '@atlaskit/icon/glyph/atlassian';
 import PackagesIcon from '@atlaskit/icon/glyph/component';
 import DocumentationIcon from '@atlaskit/icon/glyph/overview';
 import PatternsIcon from '@atlaskit/icon/glyph/issues';
 
-import {
-  RouterNavigationItem,
-  ExternalNavigationItem,
-} from './utils/linkComponents';
-import atlasKitLogo from '../../assets/atlaskit-logo.png';
 import Groups from './Groups';
 import GroupDrawer from './GroupDrawer';
 import SearchDrawer from './SearchDrawer';
 import { externalPackages as packages, docs, patterns } from '../../site';
-import type { Directory } from '../../types';
 
 import atlaskitLogo from '../../assets/atlaskit-logo-inverted.png';
 import atlaskitLogoMonochrome from '../../assets/atlaskit-logo-monochrome.png';