+++ packages/core/navigation/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/navigation
 
+## 29.3.1
+- [patch] Update typing and examples [efdbf46](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/efdbf46)
+
 ## 29.3.0
 - [minor] Add `target` prop to ResultBase component [d4a3bed](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d4a3bed)